Move and fix 404 tests
