+++ tests/Feature/Router/ErrorHandlingTest.php
@@ -9,22 +9,11 @@ use Gacela\Router\Router;
 use Gacela\Router\Routes;
 use GacelaTest\Feature\HeaderTestCase;
 use GacelaTest\Feature\Router\Fixtures\FakeController;
+use Generator;
 
 final class ErrorHandlingTest extends HeaderTestCase
 {
-    public function test_match_does_not_matches_other_methods(): void
-    {
-        $_SERVER['REQUEST_URI'] = 'https://example.org/expected/uri';
-        $_SERVER['REQUEST_METHOD'] = 'GET';
-
-        $this->expectOutputString('');
-
-        Router::configure(static function (Routes $routes): void {
-            $routes->post('expected/uri', FakeController::class, 'basicAction');
-        });
-    }
-
-    public function test_respond_status_404_when_the_uri_does_not_matches(): void
+    public function test_respond_404_status_when_uri_does_not_match(): void
     {
         $_SERVER['REQUEST_URI'] = 'https://example.org/optional/uri';
         $_SERVER['REQUEST_METHOD'] = Request::METHOD_OPTIONS;
@@ -39,5 +28,55 @@ final class ErrorHandlingTest extends HeaderTestCase
                 'response_code' => 0,
             ],
         ], $this->headers());
+    }
+
+    public function test_respond_404_status_when_method_does_not_match(): void
+    {
+        $_SERVER['REQUEST_URI'] = 'https://example.org/expected/uri';
+        $_SERVER['REQUEST_METHOD'] = 'GET';
+
+        Router::configure(static function (Routes $routes): void {
+            $routes->post('expected/uri', FakeController::class, 'basicAction');
+        });
+
+        self::assertSame([
+            [
+                'header' => 'HTTP/1.0 404 Not Found',
+                'replace' => true,
+                'response_code' => 0,
+            ],
+        ], $this->headers());
+    }
+
+    /**
+     * @dataProvider notMatchesMethodsProvider
+     */
+    public function test_respond_404_status_when_not_matches_match_methods(string $testMethod, array $givenMethods): void
+    {
+        $_SERVER['REQUEST_URI'] = 'https://example.org/expected/uri';
+        $_SERVER['REQUEST_METHOD'] = $testMethod;
+
+        Router::configure(static function (Routes $routes) use ($givenMethods): void {
+            $routes->match($givenMethods, 'expected/uri', FakeController::class, 'basicAction');
+        });
+
+        self::assertSame([
+            [
+                'header' => 'HTTP/1.0 404 Not Found',
+                'replace' => true,
+                'response_code' => 0,
+            ],
+        ], $this->headers());
+    }
+
+    public function notMatchesMethodsProvider(): Generator
+    {
+        yield [Request::METHOD_PUT, [Request::METHOD_GET, Request::METHOD_POST]];
+        yield [Request::METHOD_OPTIONS, [Request::METHOD_GET, Request::METHOD_POST]];
+        yield [Request::METHOD_GET, [Request::METHOD_PATCH, Request::METHOD_PUT, Request::METHOD_DELETE, Request::METHOD_POST]];
+        yield [Request::METHOD_CONNECT, [
+            Request::METHOD_GET, Request::METHOD_DELETE, Request::METHOD_HEAD, Request::METHOD_OPTIONS,
+            Request::METHOD_PATCH, Request::METHOD_POST, Request::METHOD_PUT, Request::METHOD_TRACE,
+        ]];
     }
 }
